+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+* Fallback to `ascii` decoder when printing help in case the locales are not properly set
+
 ## [3.1.0] - 2021-11-02
 
 ### Added